Moved firebase admin to backend
