+++ .gitignore
@@ -29,5 +29,6 @@ backend/firebase.config.js
 backend/mongo.config.js
 backend/mongo.config
 backend/firebase.config
+backend/firebaseAdmin.json
 src/Components/Admin/firebaseAdmin.js
 src/Components/Admin/firebaseAdmin.json
